Moved ModelState class to different package
