+++ blazegraph/cim2glm.bnd
@@ -1,4 +1,4 @@
-Bundle-Version: 14.0.0
+Bundle-Version: 15.0.0
 Private-Package: \
 	gov.pnnl.goss.cim2glm.components,\
 	gov.pnnl.goss.cim2glm.queryhandler.impl
