+++ blazegraph/src/main/java/gov/pnnl/goss/cim2glm/packageinfo
@@ -1,1 +1,1 @@
-version 8.0.0
+version 10.0.0
