+++ cnf/release/index.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<repository increment="1573509142382" name="Release" xmlns="http://www.osgi.org/xmlns/repository/v1.0.0">
+<repository increment="1573587549948" name="Release" xmlns="http://www.osgi.org/xmlns/repository/v1.0.0">
   <resource>
     <capability namespace="osgi.identity">
       <attribute name="osgi.identity" value="blazegraph.cim2glm"/>
@@ -661,6 +661,56 @@
       <attribute name="version" type="Version" value="9.0.0"/>
       <attribute name="bundle-symbolic-name" value="blazegraph.cim2glm"/>
       <attribute name="bundle-version" type="Version" value="14.0.0"/>
+      <directive name="uses" value="gov.pnnl.goss.cim2glm.queryhandler"/>
+    </capability>
+    <requirement namespace="osgi.wiring.package">
+      <directive name="filter" value="(&amp;(osgi.wiring.package=gov.pnnl.goss.cim2glm.queryhandler)(version&gt;=5.0.0)(!(version&gt;=6.0.0)))"/>
+    </requirement>
+    <requirement namespace="osgi.wiring.package">
+      <directive name="filter" value="(osgi.wiring.package=org.apache.commons.math3.complex)"/>
+    </requirement>
+    <requirement namespace="osgi.wiring.package">
+      <directive name="filter" value="(osgi.wiring.package=org.apache.jena.query)"/>
+    </requirement>
+    <requirement namespace="osgi.wiring.package">
+      <directive name="filter" value="(osgi.wiring.package=org.apache.jena.rdf.model)"/>
+    </requirement>
+    <requirement namespace="osgi.ee">
+      <directive name="filter" value="(&amp;(osgi.ee=JavaSE)(version=1.8))"/>
+    </requirement>
+  </resource>
+  <resource>
+    <capability namespace="osgi.identity">
+      <attribute name="osgi.identity" value="blazegraph.cim2glm"/>
+      <attribute name="type" value="osgi.bundle"/>
+      <attribute name="version" type="Version" value="15.0.0"/>
+    </capability>
+    <capability namespace="osgi.content">
+      <attribute name="osgi.content" value="d75b17838d592af815cfcc826374e1dbe87654a493e6b12884e6b7297dbb2bfd"/>
+      <attribute name="url" value="blazegraph.cim2glm/blazegraph.cim2glm-15.0.0.jar"/>
+      <attribute name="size" type="Long" value="289346"/>
+      <attribute name="mime" value="application/vnd.osgi.bundle"/>
+    </capability>
+    <capability namespace="osgi.wiring.bundle">
+      <attribute name="osgi.wiring.bundle" value="blazegraph.cim2glm"/>
+      <attribute name="bundle-version" type="Version" value="15.0.0"/>
+    </capability>
+    <capability namespace="osgi.wiring.host">
+      <attribute name="osgi.wiring.host" value="blazegraph.cim2glm"/>
+      <attribute name="bundle-version" type="Version" value="15.0.0"/>
+    </capability>
+    <capability namespace="osgi.wiring.package">
+      <attribute name="osgi.wiring.package" value="gov.pnnl.goss.cim2glm.queryhandler"/>
+      <attribute name="version" type="Version" value="5.0.1"/>
+      <attribute name="bundle-symbolic-name" value="blazegraph.cim2glm"/>
+      <attribute name="bundle-version" type="Version" value="15.0.0"/>
+      <directive name="uses" value="org.apache.jena.query"/>
+    </capability>
+    <capability namespace="osgi.wiring.package">
+      <attribute name="osgi.wiring.package" value="gov.pnnl.goss.cim2glm"/>
+      <attribute name="version" type="Version" value="8.0.0"/>
+      <attribute name="bundle-symbolic-name" value="blazegraph.cim2glm"/>
+      <attribute name="bundle-version" type="Version" value="15.0.0"/>
       <directive name="uses" value="gov.pnnl.goss.cim2glm.queryhandler"/>
     </capability>
     <requirement namespace="osgi.wiring.package">
